Split compile-fail code into two files in order to reuse most of it later
